Harden parser fallback compatibility
