--- .github/workflows/run-fix-302-codex-round.yml
@@ -1,24 +1,0 @@
-name: Apply Codex parser fallback fixes
-on:
-  push:
-    branches: [security/javascript-placeholder-bindings]
-permissions:
-  contents: write
-jobs:
-  patch:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: security/javascript-placeholder-bindings
-      - name: Apply fixes
-        shell: bash
-        run: |
-          python3 .github/fix_302_codex_round.py
-          rm -f .github/fix_302_codex_round.py .github/workflows/run-fix-302-codex-round.yml
-          git config user.name 'Ben'
-          git config user.email 'benbergen12@gmail.com'
-          git add -A
-          git commit -m 'Harden parser fallback compatibility'
-          git push origin HEAD:security/javascript-placeholder-bindings
